break out common utilities into a mixin for materialized view tests
